Update RecentView, SavedProduct

+++ src/pages/savedproduct/SavedProduct.jsx
@@ -1,46 +1,37 @@
-// Consider.jsx
-
+// components/RecentView.jsx
+import { useEffect, useState } from "react";
 import { Swiper, SwiperSlide } from 'swiper/react';
 import { Navigation, Pagination } from 'swiper/modules';
-import { StarIcon as StarSolid } from "@heroicons/react/20/solid";
 import 'swiper/css';
 import 'swiper/css/navigation';
 import 'swiper/css/pagination';
-import { Link } from 'react-router-dom';
-import React, { useState, useEffect } from 'react';
+import { Link } from "react-router-dom";
+import { X } from "lucide-react";
 
-
-function Consider({ data, review }) {
-    const [savedIds, setSavedIds] = useState([]);
-
+function SavedProduct() {
+    const [saved, setSaved] = useState([]);
+    const handleDelete = ((id) => {
+        if (window.confirm("Are you sure to delete ?")) {
+            const updated = saved.filter(p => p.ac_id !== id);
+            localStorage.setItem("saved_product", JSON.stringify(updated));
+            setSaved(updated);
+        }
+    });
 
     useEffect(() => {
-        const saved = JSON.parse(localStorage.getItem("saved_product")) || [];
-        setSavedIds(saved.map(item => item.ac_id));
+        const stored = JSON.parse(localStorage.getItem("saved_product")) || [];
+        setSaved(stored);
     }, []);
 
-
-    const handleSave = (product) => {
-        const saved = JSON.parse(localStorage.getItem("saved_product")) || [];
-        const existed = saved.find(s => s.ac_id === product.ac_id);
-
-        let updated;
-        if (existed) {
-            updated = saved.filter(s => s.ac_id !== product.ac_id);
-            alert("Removed from Saved successfully!");
-        } else {
-            updated = [...saved, product];
-            alert("Saved successfully!");
-        }
-
-        localStorage.setItem("saved_product", JSON.stringify(updated));
-        setSavedIds(updated.map(item => item.ac_id));
-    };
+    if (saved.length === 0) return (
+        <div className="noproduct h-100 flex justify-center items-center">
+            <h2 className="font-bold text-4xl text-[#DC143C]">No Product Here !</h2>
+        </div>
+    );
 
     return (
-        <div className="partners px-4 md:px-0">
-            <h2 className="text-2xl font-bold text-gray-800 mb-3">More to consider</h2>
-
+        <div className="px-4 md:px-0 my-10">
+            <h2 className="text-2xl font-bold mb-4">Saved Product</h2>
             <div className="relative w-full px-6 py-10">
 
                 {/* Custom Navigation Buttons */}
@@ -86,17 +77,10 @@ function Consider({ data, review }) {
                     }}
                     className="mySwiper"
                 >
-                    {data.map((product) => {
-                        const isSaved = savedIds.includes(product.ac_id);
-                        const productReviews = review.filter(r => r.product_id === product.ac_id);
-                        const totalReviews = productReviews.length;
-                        const avgRating = totalReviews
-                            ? (productReviews.reduce((sum, r) => sum + r.rating, 0) / totalReviews).toFixed(1)
-                            : 0;
-
-
+                    {saved.map((product) => {
                         return (
                             <SwiperSlide key={product.ac_id} className="p-6 border border-gray-200 rounded-md hover:shadow-sm">
+                                <button onClick={() => handleDelete(product.ac_id)}><X /></button>
                                 <img
                                     src={`https://storage.googleapis.com/rheemcooling/${product.brand}/${product.ac_id}/${product.ac_id}_img2.webp`}
 
@@ -110,23 +94,13 @@ function Consider({ data, review }) {
                                     {product.name}
                                 </Link>
 
-                                <div className="flex items-center gap-1 mt-1 mb-3">
-                                    <StarSolid className="w-4 h-4 text-yellow-500" />
-                                    <span className="text-sm text-gray-500">{avgRating}</span>
-                                    <span className="text-sm text-gray-500">| Sold {product.sale_quantity}</span>
-                                </div>
+
 
                                 <div className="flex gap-2 items-center mb-3">
-                                    <p className="font-bold">${product.price}</p>
+                                    <p className="font-bold text-[#DC143C]">${product.price}</p>
                                     <p className="text-sm text-gray-500 line-through">${product.old_price}</p>
                                 </div>
 
-
-                                {isSaved ? <button className="w-full bg-gray-50  hover:bg-[#DC143C] hover:text-white transition font-bold py-2 rounded-lg text-[#DC143C] border-2 border-[#DC143C] flex items-center justify-center mt-8 hover:cursor-pointer " onClick={() => handleSave(product)}>
-                                    Remove Saved
-                                </button> : <button className="w-full bg-[#DC143C] hover:bg-red-700 transition font-bold py-2 rounded-lg text-white flex items-center justify-center mt-8 hover:cursor-pointer " onClick={() => handleSave(product)}>
-                                    Save Product
-                                </button>}
                             </SwiperSlide>
                         );
                     })}
@@ -138,4 +112,4 @@ function Consider({ data, review }) {
     );
 }
 
-export default Consider;
+export default SavedProduct;
